Implement a data centering option for time series bar charts
This adds an option to time series bar charts to set whether the value
being passed to the bar should be considered the middle, start, or end
of the bar. For instance, if passing values that are set to be end
values, the bar will be drawn so that the right side of the bar aligns
with the time point on the axis instead of the center.

Author: Dylan Nugent