Fix rounding clobbering large values.
Previous behavior always rounded to 2 signifigant figures, clobbering
the difference between large values with small differences and making
their tooltips all display the same. This makes all values display to 2
decimals of precision; in the future, we may want to look at a more
dynamic option as a middle ground of fixed precision in all charts.

Author: Dylan Nugent