Allow axis labels on time series charts to be centered
Add a setting "centerAxis" for the time series labeller which will place
time labels at the center of an interval instead of the start of that
interval. For instance, with "centerAxis" true, the label for March 2004
would appear at the tick on March 15 instead of Match 1.

Author: Dylan Nugent
Reviewer: Cyril Fluck

Part of #40