[Dynamic X-Axis] Dynamic X-Axis labelling for Time Series Charts
Before, the labels along the x axis were generated using the selected
interval (months, years, quarters...). This can lead to issues where the
selected interval is not specific enough that the time series chart only
has a few labels on the x axis with not enough specificity.

This feature dynamically generates the labels along the x axis given a
maxNumberOfLabels parameter that increases the specificity until there
are enough labels along the x axis. For example, if the time range spans
a period of 2 years, the labels should not be year based as it would result
in only two labels. The specificity should be increased to be month based.

minTimeSpecificity and maxTimeSpecificity parameters were also included to
allow for a minimum and maximum time specificity to be set for  dynamic x
axis. For example, a minimum time specificity set to days means that the
labelling along the x axis cannot be more specific than the days level.

The example website was also changed so the example time series chart includes
the option to play around with the dynamic x axis.