Make labels in horizontal bar charts flush with edges
This basically reverts 485001bfb93b4341df57723eb8ba3a4bf8d9050c for horizontal bar charts, but leaves it
for pie charts. Instead of having a configurable `labelWidthMultiplier`, which
was used to determine the horizontal margin on both sides of the chart,
the left and right margins are now computed based on the longest label on
either side of the chart. This means that the labels should always be close
to flush with the edges of the chart container.

Closes #134