Fix bug #172 problem (wrong bars highlighted when mousing over bar labels in legend)
This bug didn't repro when you turn on stacked bars.
Turns out that the stacked bar code has the right behavior here
(looking up the bar's style class based on the bar label,
not just the index of the bar data object in d3's data array).

Of course, that begs the question: why must we use inline styles
to color the bars, when other aspects of the bar styling are done
through style classes? Couldn't we generate style rules
at runtime to map the bar style class to the right bar color?