Fix #172: In Vertical Bar Charts, bar groups lacking some bars wrongly color and highlight their remaining bars
`git cherry-pick 904512500ba684087519361869987331ff2a5190` from master to Addepar

Bug report: https://github.com/Addepar/ember-charts/issues/172
Refer to that for description of the bug, including problem 1 and problem 2.

For problem 1 (wrong bar colors), we create a permanent mapping
from bar labels to bar color strings, instead of using the index
of the bar datum in the bar group data array. If that array was
shorter than the total number of bar labels (that is, if that bar group
lacked some bars that other groups had), then there would be more colors
in the palette than bars in the group, and that caused the problem.
The comment in the code explains this in more detail.

For problem 2 (wrong bars being highlighted on legend mouse hover),
we use a pre-existing mapping from bar labels to bar style classes.
This mapping was already being used in the stacked bar chart case
(`stackBars = true`), which explains why this problem didn't appear
in stacked bar charts. All we have to do is use the same mapping
for unstacked bar charts too.

----------

**NEW TEST CODE DEPENDENCY ADDED:** [tinycolor.js](https://github.com/bgrins/TinyColor)

To make comparing colors with different color strings easier,
I now import the _tinycolor.js_ library for the test code.
This library has no dependencies of its own, so adding it
should be relatively low-risk, but "relatively" is the important word.
I make sure to import this library only in the "devDependencies" in Bower,
so hopefully it will not leak into the distributed module script.

----------

As an aside, it seems odd that while we use element classes to implement
bar highlighting, with one style rule for every bar with a given bar label,
bar coloring is done individually for every bar by setting the
`<rect style="...">` attribute to a style rule string.

It seems that both bar highlighting and bar coloring could be implemented
with element classes based on the bar label. We could do this by generating
style rules at runtime that map bars with a given label to a particular color.

While this change would be a good idea, it's arguably a bigger refactor
than I wanted to undertake here, so I didn't do it.

----------

Testing: all tests pass including new ones for problems 1 and 2
Code Reviewer: @BillyLittlefield
Work Item: #172
Closes #173