There was a bug with Minor Ticks 🐛 (#176)
There was a bug with Minor Ticks that arose from a complication with how D3
manages selections.  This issue only occurs in the axis is made smaller
(eliminating labels) and then made larger again.

-- COMMIT LOG  📓 --
* Lint Error fix

* Fixing a typo.

* Adding in tests to the best ability of the test container.

* Needed to move the acceptance test to a unit test.  Also had backwards logic in
the minor tick interval which was weird but not a real problem (since the
numbers were fairly consistent).

* Enriching failure messages.

* Updating Phantom JS

* Making the test more deterministic.  I'm also hoping that moving to the JS
native Date function removes some of the test breakage that I'm seeing in
travis ci.

* Release 1.1.3