Rounded Zero Percent Slice Configurations for Pie Charts (#181)
This branch fixes rounding errors and adds the missing Other Slice for when the Other Slice is incorrectly rounded to 0%. It also adds a toggle to allow users to include / exclude zero percent slices. These 0% slices respect slice ordering and number of decimal points. Finally, it contains tests for all of the above.