Fix #173 regression: Vertical Bar Charts ignored bar-specific color overrides
For each bar data point in a vertical bar chart,
you can specify the color for that bar explicitly.

Unfortunately, pull request #173 caused us to ignore these colors.
We didn't discover this bug until we tried to wire up a build
of Ember Charts with #173 to Addepar's own application.
Bad, very bad of us, and we are sorry.

This changeset fixes the regression. A new test is added to catch it.