Fix #173 regression: Vertical Bar Charts ignored bar-specific color overrides (Addepar branch version)
`git cherry-pick 611c0d5c69e3336d8d1580b43f4e5711d3251977`
from master to Addepar

For each bar data point in a vertical bar chart,
you can specify the color for that bar explicitly.
This feature should be documented in the API here:
http://opensource.addepar.com/ember-charts/#/documentation
under Vertical-Bar Options, data property, but unfortunately
the documentation is currently missing this information.

Unfortunately, pull request #173 caused us to ignore these colors.
We didn't discover this bug until we tried to wire up a build
of Ember Charts with #173 to Addepar's own application.
Bad, very bad of us, and we are sorry.

This changeset fixes the regression. A new test is added to catch it.
Closes #182
Code Reviewer: @billylittlefield