Hover labels for Line Series with Overridable Date function (#198)
We found an issue that when a REALLY small graph is generated, that the labels
for the line series can be pushed or pulled by a day.  Since we have the data
when updating the line, we can simply copy what we do for bar charts and
override the tool tip.

-- UPDATES
* Adding in a method that can be overriden by implementers that allow the nudging
of the date / time used in the hover labels.