Remove mouseover event on time-series path
Currently, time series line charts have two mouseover events that can fire -- one attached to line, and one attached to the dot that appears when hovering. The data being passed into showDetails for the line's mouseover event contains all data for the line, not a specific point. In short, this mouseover event was not doing anything except throwing errors whenever it was fired, because showDetails expects data for a single point. The correct data for a point is passed in the line marker's mouseover event, found in updateLineMarkers (in has-time-series-rule.js).

Closes #199