Remove grunt and grunt-generated dist files, update documentation (#201)
This change removes grunt dependencies and references in the documentation, as well as the actual dist global files generated from grunt. Ember-charts was converted to be an Ember-CLI addon previously, but we never removed the deprecated references and files, which this does.

We also had also used grunt for our release workflow, which therefore needs updating. The grunt plugin being used also exists as just a regular node package (https://github.com/webpro/release-it), which can perform all the same functionality and more. I've updated the documentation in the README to detail how maintainers should be releasing versions in the future, as well as added a config file for the release-it package that lets release-it handle the following functions:

* Update version numbers and commit to master
* Run ember build and push the resulting /ember-dist/ to our gh-pages branch so that the documentation site gets updated automatically
* Publish to npm

The only thing release-it isn't configured to automate is the Github versioning, which I skipped to prevent having github tokens floating around in the repo. This can be done manually.